+++ res/drawable-xxxhdpi/purple_background_gradient.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:angle="90"
+              android:centerY="0"
+              android:centerX="0.5"
+              android:startColor="@color/purple"
+              android:endColor="@color/purplelight"
+              android:type="linear"/>
+</shape>
