+++ res/layout/about.xml
@@ -7,7 +7,7 @@
     <ImageView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:id="@+id/imageView2" android:background="@drawable/purple_background" android:src="@drawable/sozysecond"/>
+            android:id="@+id/imageView2" android:background="@drawable/purple_background_gradient" android:src="@drawable/sozysecond"/>
     <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
