+++ res/layout/addprofile.xml
@@ -7,25 +7,9 @@
                   android:orientation="vertical"
                   android:layout_width="match_parent"
                   android:layout_height="match_parent" android:background="@color/white" style="@style/HoloLight">
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                  android:orientation="horizontal"
-                  android:layout_width="match_parent"
-                  android:layout_height="50dp" android:background="@color/purple" android:layout_gravity="right"
-            >
-        <Button android:layout_width="wrap_content" android:layout_height="match_parent"
-                android:background="@drawable/ic_return"
-                android:onClick="actionbarback"/>
-        <TextView android:layout_width="wrap_content" android:layout_height="match_parent"
-                  android:text="@string/addprofile" android:textSize="22sp" android:layout_gravity="center_vertical"
-                  android:gravity="center_vertical" android:textColor="@color/white" android:layout_weight="1"/>
-        <Button android:layout_width="50dp" android:layout_height="50dp"
-                android:background="@drawable/ic_ok"
-                android:onClick="ok"
-                android:layout_weight="0"/>
-    </LinearLayout>
 
 
-    <TextView android:layout_width="match_parent" android:layout_height="wrap_content"
+        <TextView android:layout_width="match_parent" android:layout_height="wrap_content"
             android:text="@string/name"
             android:id="@+id/namehead"
             android:textColor="@color/black"
