+++ res/layout/ftppathsselection.xml
@@ -7,21 +7,6 @@
                   android:orientation="vertical"
                   android:layout_width="match_parent"
                   android:layout_height="match_parent" android:background="@color/white">
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                  android:orientation="horizontal"
-                  android:layout_width="match_parent"
-                  android:layout_height="50dp" android:background="@color/purple" android:layout_gravity="right">
-        <Button android:layout_width="wrap_content" android:layout_height="match_parent"
-                android:background="@drawable/ic_return"
-                android:onClick="actionbarback"/>
-        <TextView android:layout_width="wrap_content" android:layout_height="match_parent"
-                  android:text="@string/choosepath" android:textSize="22dp" android:layout_gravity="center_vertical"
-                  android:gravity="center_vertical" android:textColor="@color/white" android:layout_weight="2"/>
-        <Button android:layout_width="50dp" android:layout_height="50dp"
-                android:background="@drawable/ic_ok"
-                android:onClick="ok"
-                android:layout_weight="0" android:id="@+id/ftpokbutton"/>
-    </LinearLayout>
 
 
     <TextView android:layout_width="match_parent" android:layout_height="wrap_content"
