+++ res/layout/listitemprofiles.xml
@@ -13,7 +13,7 @@
     <TextView android:layout_width="match_parent" android:layout_height="wrap_content"
             android:id="@+id/profilenameitem"
             android:textColor="@color/black"
-            android:textSize="25dp"/>
+            android:textSize="25sp"/>
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
                   android:orientation="horizontal"
                   android:layout_width="match_parent"
