+++ res/layout/profile_info.xml
@@ -1,0 +1,90 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:orientation="vertical"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent">
+
+    <LinearLayout
+            android:orientation="vertical"
+            android:layout_width="fill_parent"
+            android:layout_height="fill_parent" android:background="@color/white"
+            android:paddingLeft="10dp" android:paddingTop="5dp" android:paddingRight="10dp">
+
+        <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="New Text"
+                android:id="@+id/info_name" android:background="@color/white"
+                android:layout_gravity="center_horizontal"
+                android:textSize="25sp"
+                android:textStyle="bold"
+                android:textColor="@color/black"/>
+        <TextView android:layout_width="match_parent" android:layout_height="2dp"
+                  android:background="@color/purplelight"/>
+        <TextView
+                android:layout_width="match_parent"
+                android:layout_height="20dp"
+                android:id="@+id/textView4"/>
+        <LinearLayout
+                android:orientation="horizontal"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" android:layout_gravity="left">
+            <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/addrhead"
+                    android:layout_gravity="left" android:background="@color/white"
+                    android:textStyle="bold" android:textColor="@color/black" android:textSize="15sp"/>
+            <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/addrhead"
+                    android:id="@+id/info_address"
+                    android:layout_gravity="left" android:background="@color/white"
+                    android:textStyle="normal" android:textColor="@color/black" android:textSize="15sp"
+                    android:paddingLeft="5dp"/>
+        </LinearLayout>
+        <LinearLayout
+                android:orientation="horizontal"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/userhead" android:layout_gravity="left" android:background="@color/white"
+                      android:textStyle="bold" android:textColor="@color/black" android:textSize="15sp"/>
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/userhead" android:id="@+id/info_user" android:layout_gravity="left"
+                      android:background="@color/white" android:textStyle="normal" android:textColor="@color/black"
+                      android:textSize="15sp" android:paddingLeft="5dp"/>
+        </LinearLayout>
+        <LinearLayout android:orientation="horizontal" android:layout_width="match_parent"
+                      android:layout_height="wrap_content">
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/localpathhead" android:layout_gravity="left" android:background="@color/white"
+                      android:textStyle="bold" android:textColor="@color/black" android:textSize="15sp"/>
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/localpathhead" android:id="@+id/info_localpath" android:layout_gravity="left"
+                      android:background="@color/white" android:textStyle="normal" android:textColor="@color/black"
+                      android:textSize="15sp" android:paddingLeft="5dp"/>
+        </LinearLayout>
+        <LinearLayout android:orientation="horizontal" android:layout_width="match_parent"
+                      android:layout_height="wrap_content">
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/remotepathhead" android:layout_gravity="left" android:background="@color/white"
+                      android:textStyle="bold" android:textColor="@color/black" android:textSize="15sp"/>
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/remotepathhead" android:id="@+id/info_remotepath" android:layout_gravity="left"
+                      android:background="@color/white" android:textStyle="normal" android:textColor="@color/black"
+                      android:textSize="15sp" android:paddingLeft="5dp"/>
+        </LinearLayout>
+        <LinearLayout android:orientation="horizontal" android:layout_width="match_parent"
+                      android:layout_height="wrap_content">
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/typehead" android:layout_gravity="left" android:background="@color/white"
+                      android:textStyle="bold" android:textColor="@color/black" android:textSize="15sp"/>
+            <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
+                      android:text="@string/typehead" android:id="@+id/info_type" android:layout_gravity="left"
+                      android:background="@color/white" android:textStyle="normal" android:textColor="@color/black"
+                      android:textSize="15sp" android:paddingLeft="5dp"/>
+        </LinearLayout>
+    </LinearLayout>
+</LinearLayout>
