+++ res/menu/add_profile_actions.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/ok_action"
+          android:icon="@drawable/ic_ok"
+          android:showAsAction="always"
+          android:title="@string/savebutton"/>
+</menu>
