+++ res/menu/info_actions.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:title="@string/delete"
+          android:icon="@drawable/ic_delete"
+          android:id="@+id/deletemenu"
+            android:showAsAction="always"/>
+    <item android:title="@string/start"
+          android:id="@+id/startmenu"
+          android:showAsAction="always"
+          android:icon="@drawable/ic_start"/>
+</menu>
