+++ src/com/soutvoid/ProjectSozy/PathsSelection.java
@@ -7,6 +7,8 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.preference.DialogPreference;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.*;
 import android.content.DialogInterface;
@@ -25,7 +27,6 @@ public class PathsSelection extends ListActivity {
     private List<String> directoryEntriesShow = new ArrayList<String>();
     private File currentDirectory = new File("/");
     private File currentFile = new File("");
-    private Button ok;
 
 
     //заполняем лист содержимым папки
@@ -67,7 +68,12 @@ public class PathsSelection extends ListActivity {
                 public void onClick(DialogInterface dialog, int which) {
                     //дествия по нажатию да
                     currentFile = aDirectory;
-                    ok(ok);
+                    Intent i = new Intent();
+                    if (currentFile.getName() == "")
+                        i.putExtra("path", currentDirectory.getAbsolutePath());
+                    else i.putExtra("path", currentFile.getAbsolutePath());
+                    setResult(RESULT_OK, i);
+                    finish();
 
                 }
             };
@@ -111,22 +117,33 @@ public class PathsSelection extends ListActivity {
     public void onCreate(Bundle icicle) {
         super.onCreate(icicle);
         setContentView(R.layout.pathsselection);
-
-        ok = (Button)findViewById(R.id.okbutton);
+        getActionBar().setIcon(R.drawable.ic_action_back);
+        getActionBar().setHomeButtonEnabled(true);
 
         browseTo(new File("/sdcard"));
     }
 
-    public void actionbarback(View view) {
-        onBackPressed();
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.add_profile_actions, menu);
+        return true;
     }
 
-    public void ok(View view) {
-        Intent i = new Intent();
-        if (currentFile.getName() == "")
-        i.putExtra("path", currentDirectory.getAbsolutePath());
-        else i.putExtra("path", currentFile.getAbsolutePath());
-        setResult(RESULT_OK, i);
-        finish();
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                onBackPressed();
+                break;
+            case R.id.ok_action:
+                Intent i = new Intent();
+                if (currentFile.getName() == "")
+                    i.putExtra("path", currentDirectory.getAbsolutePath());
+                else i.putExtra("path", currentFile.getAbsolutePath());
+                setResult(RESULT_OK, i);
+                finish();
+                break;
+        }
+        return true;
     }
 }
